[🎸기타] axios 파일 경로 변경
