+++ src/axios/case28/audio/audio_regression.tsx
@@ -1,22 +1,22 @@
+/* 피아노 악보 생성 - 음성 회귀 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-import base64DataToFile from './base64DataToFile'
+import base64DataToFile from '../../base64DataToFile'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const videoRegression = () => {
+const audioRegression = () => {
   let data = {
-    file: base64DataToFile(detailData, '이미지이름', 'image/gif'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/a937e/inference',
+    file: base64DataToFile(detailData, '오디오이름', 'audio/midi'),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/29683/inference',
   }
 
   setLoading(true)
 
   axios
     .post('/inference/file_req_ajx', data, {
-      // processData, contentType, dataType은 Axios에서 사용되지 않는 속성이다
       headers: {
         'Content-Type': 'multipart/form-data',
       },
@@ -25,12 +25,13 @@ const videoRegression = () => {
     .then(res => {
       let json = res.data
       if (json.res == 'true') {
-        var response_data = json.response.data
+        let response_data = json.response.data
         if (response_data == null) {
           response_data = json.response.inference
         }
-        // $('#resImgSrc').attr('src', 'data:image/jpg;base64,' + response_data)
-        // $('div.inner_next').addClass('show_img')
+        // 결과 들어가는 부분
+        // $("#tarMidi").attr("src", "data:audio/midi;base64," + response_data);
+        // $(".inner_next").css('background', 'none');
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -43,4 +44,4 @@ const videoRegression = () => {
     })
 }
 
-export default videoRegression
+export default audioRegression
