+++ src/axios/case28/satellite/satellite_clustering.tsx
@@ -1,19 +1,20 @@
-/* 악성코드 이상탐지 - 바이너리 이상탐지 */
+/* 지표면 위성 사진 군집화 - 위성 군집화 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-import base64DataToFile from './base64DataToFile'
+import base64DataToFile from '../../base64DataToFile'
+import { Keyword } from '@/components'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const binaryClassification = () => {
+const satelliteClustering = () => {
   let data = {
     file: base64DataToFile(detailData, '사진이름', 'image/png'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/09a5a/inference',
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/ff463/inference',
   }
-
   setLoading(true)
+  let class_info: any = { forested: '산림화', desertified: '사막화' }
 
   axios
     .post('/inference/file_req_ajx', data, {
@@ -29,15 +30,8 @@ const binaryClassification = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
-        /* 결과 */
-        if (response_data == 'benign') {
-          // 정상 결과 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_pass");
-        } else {
-          console.log('파손')
-          // 파손 결과 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_nonpass");
-        }
+        response_data = class_info[response_data]
+        ;<Keyword option={1} label={response_data} />
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -50,4 +44,4 @@ const binaryClassification = () => {
     })
 }
 
-export default binaryClassification
+export default satelliteClustering
