+++ src/axios/case28/video/video_anomaly.tsx
@@ -1,22 +1,22 @@
-/* 악성코드 분류 - 바이너리 분류 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-import base64DataToFile from './base64DataToFile'
+import base64DataToFile from '../../base64DataToFile'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const binaryAnomaly = () => {
+const videoAnomaly = () => {
   let data = {
-    file: base64DataToFile(detailData, '사진이름', 'image/png'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/58d37/inference',
+    file: base64DataToFile(detailData, '이미지이름', 'image/gif'),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/90924/inference',
   }
 
   setLoading(true)
 
   axios
     .post('/inference/file_req_ajx', data, {
+      // processData, contentType, dataType은 Axios에서 사용되지 않는 속성이다
       headers: {
         'Content-Type': 'multipart/form-data',
       },
@@ -25,20 +25,17 @@ const binaryAnomaly = () => {
     .then(res => {
       let json = res.data
       if (json.res == 'true') {
-        let response_data = json.response.data
+        var response_data = json.response.data
         if (response_data == null) {
           response_data = json.response.inference
         }
-        /* 결과 */
-        if (response_data == 'benign') {
-          // 양성 데이터 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_pass");
-        } else if (response_data == 'malware') {
-          // 악성 데이터 들어가는 부분
-          // $("div.wrap_next").addClass("show_alert_nonpass");
+        if (response_data == 'normal') {
+          // $('div.wrap_next').addClass('show_alert_pass')
         } else {
-          alert('API 호출에 실패했습니다.')
+          // $('div.wrap_next').addClass('show_alert_nonpass')
         }
+      } else {
+        alert('API 호출에 실패했습니다.')
       }
     })
     .catch(err => {
@@ -49,4 +46,4 @@ const binaryAnomaly = () => {
     })
 }
 
-export default binaryAnomaly
+export default videoAnomaly
