+++ src/axios/school/korea/ku_1200.tsx
