+++ src/axios/school/korea/ku_1201.tsx
