+++ src/axios/school/korea/ku_1202.tsx
