+++ src/axios/school/korea/ku_1203.tsx
