+++ src/axios/school/korea/ku_1204.tsx
@@ -1,17 +1,16 @@
-/* 음성단어분류 - 음성 분류 */
+/* 고려대학교 - 가짜 얼굴 판별 서비스 1204 */
 import axios from 'axios'
 import { detailDataAtom, loadingAtom } from '@/atoms'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
-import base64DataToFile from './base64DataToFile'
-import { Keyword } from '@/components';
+import base64DataToFile from '../../base64DataToFile'
 
 const detailData = useRecoilValue<any>(detailDataAtom)
 const setLoading = useSetRecoilState(loadingAtom)
 
-const audioClassification = () => {
+const ku1204 = () => {
   let data = {
-    file: base64DataToFile(detailData, '오디오이름', 'audio/wav'),
-    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/c68e0/inference',
+    file: base64DataToFile(detailData, '사진이름', 'image/jpeg'),
+    url: 'http://dl.idro3vub.aica.t3q.ai/model/api/d29f6/inference',
   }
 
   setLoading(true)
@@ -30,10 +29,13 @@ const audioClassification = () => {
         if (response_data == null) {
           response_data = json.response.inference
         }
-        // 결과 들어가는 부분
-        // $(".result_alert").html(response_data);
-        // $("div.wrap_next").addClass("show_alert_pass");
-        <Keyword option={1} label={response_data}/>
+        if (response_data == 'natural image') {
+          // 정상
+          // $('div.wrap_next').addClass('show_alert_pass')
+        } else if (response_data == 'fake image') {
+          // 비정상
+          // $('div.wrap_next').addClass('show_alert_nonpass')
+        }
       } else {
         alert('API 호출에 실패했습니다.')
       }
@@ -46,4 +48,4 @@ const audioClassification = () => {
     })
 }
 
-export default audioClassification
+export default ku1204
