+++ src/axios/school/korea/ku_1205.tsx
