+++ src/axios/school/kyungpook/knu_1101.tsx
