+++ src/axios/school/kyungpook/knu_1102.tsx
